matrix: Transfer ownership of client to pointer array
The pointer array uses a `g_object_unref()` as cleanup func so we need
to add a ref'ed object to the array as `client` is marked as
`g_autoptr`.

This fixes a crash on shutdown in eignvalue when we unref the client we
get from the liststore returned by `cm_matrix_get_clients_list()` via
`g_list_model_get_item()` (which gives us a reference).