secret-store: Get rid of old chatty account handling
When this code still lived in chatty the labels of secrets
were of the form "Chatty password for <username>".

This changed when chatty switch to using libcmatrix,
with 0.7.0~rc0 in October 2022 (and 0.7.0 in February 2023).

It's about time to get rid of this migration layer:
Greedily grabbing every secret matching the schema is a recipe for confusion
when there can be multiple applications on the same system using libcmatrix.