build: Build shared library too
This will be needed for users of introspection but is also sometimes
nice for development purposes.

We set a soversion of `0u` (u for unstable ABI) so that we have `0` for
the first ABI stable release.

We need to make gio a `Requires:` for that as we expose e.g.
`GListModel`.