room: Ensure list of joined members is current when getting member list
When `cm_room_get_joined_members()` is called, the callee has
a (reasonable) expectation that the list of room members is complete.
Ensure it is complete by querying the server with
`cm_load_joined_members_*()`.

`cm_load_joined_members_*()` was only called for encrypted chats
in `ensure_encryption()` leaving unencrypted chats with incomplete
room members.

The codepath in `ensure_encryption()` is left unchanged as it
- always needs the member list to fetch device keys
- checks whether the member list is loaded/loading already
  (so we're not wasting cycles on a pending/completed member loading
   due to the newly introduced callsite)

Fixes: https://source.puri.sm/Librem5/libcmatrix/-/issues/56

Signed-off-by: Evangelos Ribeiro Tzaras <devrtz@fortysixandtwo.eu>